requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
discord.py==1.4.1
tinydb==4.1.1
xboxapi==2.0.1
yarl==1.6.0
translate
cutlet
unidic-lite
